app/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
django-cors-headers
djangorestframework
djangorestframework_simplejwt
django-environ
celery==5.0.5
gunicorn==19.9.0
django-phonenumber-field[phonenumbers]
Jinja2
smsapi-client